requirements.txt
kivy==2.3.0
kivymd==1.2.0
watchdog
folium
matplotlib==3.5.0
kivy-garden.mapview
plyer==2.1.0
bs4
firebase-admin
pandas
bs4
requests
schedule